+++ CHANGELOG.md
@@ -34,6 +34,17 @@ document.
 - The plug-in used to temporarily store the refunded custom fee amounts entered 
   in the Admin panel during credit memo creation has been removed
 
+## [1.3.1]
+
+### Changed
+
+- Totals in the Custom Order Fees Report are now sorted by order currency
+
+### Fixed
+
+- Custom fee totals were rendered for the related order instead of the 
+  invoice or credit memo on sales PDF files printed from the Magento Admin panel 
+
 ## [1.3.0]
 
 ### Added
@@ -153,7 +164,8 @@ document.
 [Keep a Changelog]: https://keepachangelog.com/en/1.1.0
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 [README]: ./README.md
-[Unreleased]: https://github.com/JosephLeedy/magento2-module-custom-fees/compare/1.3.0...HEAD
+[Unreleased]: https://github.com/JosephLeedy/magento2-module-custom-fees/compare/1.3.1...HEAD
+[1.3.1]: https://github.com/JosephLeedy/magento2-module-custom-fees/releases/tag/1.3.1
 [1.3.0]: https://github.com/JosephLeedy/magento2-module-custom-fees/releases/tag/1.3.0
 [1.2.3]: https://github.com/JosephLeedy/magento2-module-custom-fees/releases/tag/1.2.3
 [1.2.2]: https://github.com/JosephLeedy/magento2-module-custom-fees/releases/tag/1.2.2